Components: Unify forms
add prepare-from for remove same components (like treasury-increase-form.vue, savings-deposit-form.vue etc.)
add review-form for the same purpose
start remove same components
