+++ src/components/forms/types.ts
@@ -1,0 +1,1 @@
+export type INPUT_MODE = 'NATIVE' | 'TOKEN';
