+++ src/components/savings/savings-empty.vue
@@ -36,7 +36,6 @@ import { formatPercents, formatToNative } from '@/utils/format';
 import ActionButton from '@/components/buttons/action-button.vue';
 import { BarChart } from '@/components/charts';
 import { SecondaryPage } from '@/components/layout';
-import { toggleSingleItem } from '@/components/toggle/toggle-root';
 
 export default Vue.extend({
   name: 'SavingsEmpty',
